Caser搞好了，准备把其他几个sequential model搞好，并把policy嫁接在上面

+++ src/core/static_dataset.py
@@ -47,8 +47,8 @@ class StaticDataset(Dataset):
         else:
             self.score = score
 
-        self.x_numpy = self.x_numpy.astype(np.float)
-        self.y_numpy = self.y_numpy.astype(np.float)
+        self.x_numpy = self.x_numpy.astype(float)
+        self.y_numpy = self.y_numpy.astype(float)
 
         self.len = len(self.x_numpy)
